// types/Shipping.ts
export type ShippingT = {
  id?: number; //6
  name_surname: string;
  user_id?: number; //1;
  address: string; //"Rakowiecka 45/4";
  city: string; //"Warszawa";
  state?: string; //"";
  zip_code: string; //"27-136";
  country: string; //"Poland";
  phone: string; //"+48 123456789";
  email: string; //"axios.programming@gmail.com";
  is_default?: boolean; //false;
  created_at?: string; //"2023-01-29T17:25:56.705828Z";
  updated_at?: string; //"2023-01-29T17:25:56.705828Z";
};